ajira_guild/requirements.txt
Django
djangorestframework
mysqlclient
django-crispy-forms
python-dotenv
Pillow
django-filter
django-model-utils
django-mysql
